+++ fiboa_cli/const.py
@@ -10,3 +10,14 @@ SUPPORTED_PROTOCOLS = ["http", "https", "s3", "gs"]
 STAC_COLLECTION_SCHEMA = "http://schemas.stacspec.org/v{version}/collection-spec/json-schema/collection.json"
 GEOPARQUET_SCHEMA = "https://geoparquet.org/releases/v{version}/schema.json"
 STAC_TABLE_EXTENSION = "https://stac-extensions.github.io/table/v1.2.0/schema.json"
+
+COMPRESSION_METHODS = ["brotli", "gzip", "lz4", "snappy", "zstd", "none"]
+
+CORE_COLUMNS = [
+    "id",
+    "geometry",
+    "area",
+    "perimeter",
+    "determination_datetime",
+    "determination_method",
+]