+++ fiboa_cli/datasets/be_wa.py
@@ -91,7 +91,6 @@ def convert(output_file, cache = None, **kwargs):
         license=LICENSE,
         layer_filter=lambda layer, uri: layer == LAYER,
         file_migration=file_migration,
-        explode_multipolygon=True,
         index_as_id=True,
         **kwargs
     )