+++ fiboa_cli/merge.py
@@ -2,18 +2,11 @@ import os
 
 import pandas as pd
 
+from .const import CORE_COLUMNS
 from .parquet import create_parquet
-from .util import load_parquet_data, load_parquet_schema, log, parse_metadata
+from .util import load_parquet_data, load_parquet_schema, log, merge_schemas, parse_metadata, pick_schemas, is_schema_empty
 from .version import fiboa_version
 
-DEFAULT_COLUMNS = [
-    "id",
-    "geometry",
-    "area",
-    "perimeter",
-    "determination_datetime",
-    "determination_method",
-]
 DEFAULT_CRS = "EPSG:4326"
 
 def merge(datasets, out, crs = DEFAULT_CRS, includes = [], excludes = [], extensions = [], compression = None, geoparquet1 = False):
@@ -21,15 +14,17 @@ def merge(datasets, out, crs = DEFAULT_CRS, includes = [], excludes = [], extens
     if dir:
         os.makedirs(dir, exist_ok=True)
 
-    columns = DEFAULT_COLUMNS.copy()
+    columns = CORE_COLUMNS.copy()
     columns.extend(includes)
     columns = list(set(columns) - set(excludes))
 
     # Load the datasets
     all_gdf = []
+    custom_schemas = {}
     for dataset in datasets:
         # Load the dataset
         schema = load_parquet_schema(dataset)
+        collection = parse_metadata(schema, b"fiboa")
         file_columns = list(set(columns) & set(schema.names))
         gdf = load_parquet_data(dataset, columns=file_columns)
 
@@ -37,11 +32,14 @@ def merge(datasets, out, crs = DEFAULT_CRS, includes = [], excludes = [], extens
         gdf.to_crs(crs=crs, inplace=True)
 
         # Add collection column to each dataset
-        collection = parse_metadata(schema, b"fiboa")
         if collection is not None and "id" in collection:
             gdf["collection"] = collection["id"]
         else:
             gdf["collection"] = os.path.splitext(os.path.basename(dataset))[0]
+
+        # Merge custom schemas
+        custom_schema = collection.get("fiboa_custom_schemas", {})
+        custom_schemas = merge_schemas(custom_schemas, custom_schema)
 
         all_gdf.append(gdf)
 
@@ -57,6 +55,11 @@ def merge(datasets, out, crs = DEFAULT_CRS, includes = [], excludes = [], extens
         "fiboa_version": fiboa_version,
         "fiboa_extensions": extensions,
     }
+
+    # Add custom schemas
+    custom_schemas = pick_schemas(custom_schemas, columns)
+    if not is_schema_empty(custom_schemas):
+        collection["fiboa_custom_schemas"] = custom_schemas
 
     # Write the merged dataset to the output file
     create_parquet(merged, columns, collection, out, {}, compression=compression, geoparquet1=geoparquet1)