+++ fiboa_cli/util.py
@@ -224,6 +224,22 @@ def parse_converter_input_files(ctx, param, value):
     return sources
 
 
+def parse_map(value, separator = "="):
+    if value is None:
+        return {}
+    elif not isinstance(value, tuple):
+        raise click.BadParameter('Input files must be a tuple')
+    elif len(value) == 0:
+         return {}
+
+    mapping = {}
+    for v in value:
+        key, value = v.split(separator, 2)
+        mapping[key] = value
+
+    return mapping
+
+
 def name_from_uri(url):
     if "://" in url:
         try:
@@ -246,6 +262,10 @@ def check_ext_schema_for_cli(value, allow_none = False):
     return map_
 
 
+def is_schema_empty(schema):
+    return len(schema.get("properties", {})) == 0 and len(schema.get("required", {})) == 0
+
+
 def merge_schemas(*schemas):
     """Merge multiple schemas into one"""
     result = {
@@ -256,6 +276,24 @@ def merge_schemas(*schemas):
         schema = migrate_schema(schema)
         result["required"] += schema.get("required", [])
         result["properties"].update(schema.get("properties", {}))
+
+    return result
+
+
+def pick_schemas(schema, property_names, rename = {}):
+    """Pick and rename schemas for specific properties"""
+    result = {
+        "required": [],
+        "properties": {}
+    }
+    required = schema.get("required", [])
+    properties = schema.get("properties", {})
+    for prop in property_names:
+        prop2 = rename[prop] if prop in rename else prop
+        if prop in required:
+            result["required"].append(prop2)
+        if prop in properties:
+            result["properties"][prop2] = properties[prop]
 
     return result